+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Arrayx 增加 indices, contains, indexOf
 * 增加 Throwable
 * File 增加 listRecursively
+* Number 增加 toByte、toShort、toInt 等方法
 
 ## 0.6.1
 * Arrayx 增加 plus, contentToString